requirements.txt
requests
qrcode
beautifulsoup4
pillow

